changes paths in feature extractor pipelines
