+++ bacpipe/embedding_generation_pipelines/feature_extractors/animal2vec_mk.py
@@ -1,7 +1,7 @@
 from fairseq import checkpoint_utils
 import numpy as np
 import torch
-from bacpipe.model_utils.animal2vec_nn.nn import chunk_and_normalize
+from bacpipe.model_specific_utils.animal2vec_nn.nn import chunk_and_normalize
 from ..utils import ModelBaseClass
 
 SAMPLE_RATE = 8000
