+++ bacpipe/model_specific_utils/perch_chirp/chirp/audio_utils.py
@@ -26,8 +26,8 @@ import tempfile
 from typing import Callable, Generator, Sequence
 import warnings
 
-# from bacpipe.model_utils.perch_chirp.chirp import path_utils
-from bacpipe.model_utils.perch_chirp.chirp import signal
+# from bacpipe.model_specific_utils.perch_chirp.chirp import path_utils
+from bacpipe.model_specific_utils.perch_chirp.chirp import signal
 from etils import epath
 from jax import lax
 from jax import numpy as jnp
