+++ bacpipe/model_specific_utils/perch_chirp/chirp/preprocessing/pipeline.py
@@ -19,10 +19,10 @@ import dataclasses
 from typing import Any, Iterable, Sequence, Tuple
 
 from absl import logging
-from bacpipe.model_utils.perch_chirp.chirp import audio_utils
-from bacpipe.model_utils.perch_chirp.chirp.models import frontend
-from bacpipe.model_utils.perch_chirp.chirp.taxonomy import namespace
-from bacpipe.model_utils.perch_chirp.chirp.taxonomy import namespace_db
+from bacpipe.model_specific_utils.perch_chirp.chirp import audio_utils
+from bacpipe.model_specific_utils.perch_chirp.chirp.models import frontend
+from bacpipe.model_specific_utils.perch_chirp.chirp.taxonomy import namespace
+from bacpipe.model_specific_utils.perch_chirp.chirp.taxonomy import namespace_db
 import jax
 from jax import numpy as jnp
 import pandas as pd
