+++ bacpipe/model_specific_utils/perch_chirp/chirp/projects/zoo/taxonomy_model_tf.py
@@ -19,8 +19,8 @@ import dataclasses
 from typing import Any
 
 from absl import logging
-from bacpipe.model_utils.perch_chirp.chirp.projects.zoo import zoo_interface
-from bacpipe.model_utils.perch_chirp.chirp.taxonomy import namespace
+from bacpipe.model_specific_utils.perch_chirp.chirp.projects.zoo import zoo_interface
+from bacpipe.model_specific_utils.perch_chirp.chirp.taxonomy import namespace
 from etils import epath
 from ml_collections import config_dict
 import numpy as np
