+++ bacpipe/model_specific_utils/perch_chirp/chirp/taxonomy/namespace_db.py
@@ -20,8 +20,8 @@ import json
 import os
 import typing
 
-from bacpipe.model_utils.perch_chirp.chirp import path_utils
-from bacpipe.model_utils.perch_chirp.chirp.taxonomy import namespace
+from bacpipe.model_specific_utils.perch_chirp.chirp import path_utils
+from bacpipe.model_specific_utils.perch_chirp.chirp.taxonomy import namespace
 from etils import epath
 
 TAXONOMY_DATABASE_FILENAME = "taxonomy/taxonomy_database.json"
